+++ src/GUI/FloatingButton/Hooks.cpp
@@ -4,8 +4,6 @@
 
 bool FloatingMenuLayer::init()
 {
-    return MenuLayer::init();
-
     FloatingUIManager::get()->updateSprites();
 
     MenuLayer::init();